requirements.txt
frappe
erpnext
zenpy
phonenumberslite